+++ source/extensions/omni.flux.utils.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.25.1]
+### Changed
+- Sorted `__all__` exports alphabetically
+
 ## [1.25.0]
 ### Added
 - Added NicknameField to handle nickname attribute changes